Switch to unlicense for this repo
My only worry is if people clone this and don't change this license, they might be licensing their software as public domain without realizing it. On the flipside, implying that our exceedingly simple template is GPL may imply limitations to their rights to their own software if they start by cloning this. I'm not sure the best way forward, but I feel like public domaining this repo is a good idea.